# BUG: Timestamp.replace chaining not compat with datetime.replace