- [x] closes #15683
- [x] tests added / passed
- [x] passes `git diff upstream/master -u -- "*.py" | flake8 --diff`
- [x] whatsnew entry

This is a clean version of https://github.com/pandas-dev/pandas/pull/16110 and the last thing I’m going to do with this issue.