@gfyoung 2017-08-28
@sscherfke : This look pretty good save for a minor linting issue, which you can see here:

https://travis-ci.org/pandas-dev/pandas/jobs/269100110#L1809