@sscherfke 2017-09-04
Can this PR *please* be merge? This function computes **wrong results** since more than half a year by now!