@TomAugspurger 2017-09-04
From https://github.com/pandas-dev/pandas/pull/16110#discussion_r133924066

> `change the definition of tz_convert_single to cpdef int64_t`

Can you try that and check if it improves performance at all?